Add "tcpfd" protocol for inherited listeners
Add a new `types.TCPFD` `TransportProtocol` that allows forwarding
connections from an already-open listener file descriptor.

This allows `gvproxy` (or an in-process `virtualnetwork`) to handle
connections without opening the socket itself.

For example when using socket activation with `systemd` (on Linux):
https://www.man7.org/linux/man-pages/man5/systemd.socket.5.html
https://www.man7.org/linux/man-pages/man3/sd_listen_fds.3.html

Or `launchd` (on Darwin/macOS):
https://keith.github.io/xcode-man-pages/launchd.plist.5.html#Sockets
https://developer.apple.com/documentation/xpc/1505523-launch_activate_socket

The init system waits for an incoming connection, launches a service,
and then hands the already-open listner socket to the new service
allowing it to be started on-demand.

With this change that connection can be forwarded directly to the guest
network.

The caller retains ownership of the file descriptor that's being
Exposed.  This way there's no ambiguity as to the state of the file
descriptor; success or failure act the same.

Signed-off-by: Chris Pick <chris@chrispick.com>